requirements.txt
transformers
torch
torchvision
chromadb
streamlit
google-generativeai
lightkurve
pillow
matplotlib
requests
numpy
